Add a core data manager
Note: This provides a singelton for an in-memory data store instead of
using the default SQLite database as the storage. Access to this core
data manager will be via the singelton:

  CoreDataManager.inMemory

If a future release requires access to the an SQLite version, this will
need a new singleton to be defined by adding a computed property using
something like:

  static var shared: CoreDataManager = CoreDataManager()

This will then need to replace any use of CoreDataManager.inMemory with
CoreDataManager.shared.