Add core data manager to content view
This is passed as a parameter to the initialiser so a different manager
can be used between the actual application and the preview in the
future. That is, for the real application we can use a SQLite memory
store and for the preview we can use the in-memory store.

This not currently implemented during initial development since it is
easier to prototype using an in-memory store. Later we can add the
SQLite data store if required.