Refactor to move fetch request to Entity
It is preferred to have the fetch request defined as part of the
entity. In this case we just create a property on the entity to return
a fetch request. This is different to the default Core Data
implementation (when using auto generated entities) which defines a
method on the entity to return a fetch request.

Note: This is a class property, which must be accessed using
Move.fetchRequest.

This is preferred since creating a fetch request uses a string to
determine the entity name, which could lead to errors if the entity
name is not known outside the class. Also, it keeps the fetch request
as part of the class, rather than having to create it as a seperate
variable - as was done before this refactoring.