Refresh the list when navigating back from detail
When navigating back from the detail view the list was not being
updated with any modifications made to a movie name in the detail view.
This is a hack to force an update to the List when a movie name is
changed.

A state variable is passed to the detail view and set to true to
indicate the content view list needs to be refreshed. In order to have
the list refreshed, we have to use this new boolean somewhere. Here it
is used to set the accent colour of the list.